Implemented project following Clean Architecture
