+++ run-tests.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Run test with coverage
+sbt clean coverage test coverageReport
