--- src/main/scala/com/igobrilhante/GHSystem.scala
@@ -1,11 +1,0 @@
-package com.igobrilhante
-
-import akka.actor.typed.Behavior
-import akka.actor.typed.scaladsl.Behaviors
-
-object GHSystem {
-
-  def apply(): Behavior[GHSystem.Command] = Behaviors.empty
-
-  trait Command
-}
