+++ src/main/scala/com/igobrilhante/github/application/algorithms/RankAlgorithm.scala
@@ -1,0 +1,9 @@
+package com.igobrilhante.github.application.algorithms
+
+import com.igobrilhante.github.domain.entities.GHContributor
+
+trait RankAlgorithm[F[_]] {
+
+  def computeRanking(organizationId: String): F[Option[List[GHContributor]]]
+
+}
