+++ src/main/scala/com/igobrilhante/github/application/commons/Logging.scala
@@ -1,0 +1,7 @@
+package com.igobrilhante.github.application.commons
+
+import org.slf4j.{Logger, LoggerFactory}
+
+trait Logging {
+  val logger: Logger = LoggerFactory.getLogger(this.getClass)
+}
