+++ src/main/scala/com/igobrilhante/github/application/usecases/UseCase.scala
@@ -1,0 +1,11 @@
+package com.igobrilhante.github.application.usecases
+
+import scala.concurrent.Future
+
+trait UseCase[F[_], InputData, OutputData] {
+
+  def execute(inputData: InputData): F[OutputData]
+
+}
+
+trait UseCaseFuture[I, O] extends UseCase[Future, I, O]
