--- src/main/scala/com/igobrilhante/github/commons/CacheHelper.scala
@@ -1,14 +1,0 @@
-package com.igobrilhante.github.commons
-
-import net.sf.ehcache.{Cache, CacheManager}
-
-object CacheHelper {
-
-  val GithubContributorsCache = "githubContributorsCache"
-  private val cacheManager = new CacheManager
-
-  cacheManager.addCache(GithubContributorsCache)
-
-  def getContributorsCache: Cache = cacheManager.getCache(GithubContributorsCache)
-
-}
