--- src/main/scala/com/igobrilhante/github/commons/Logging.scala
@@ -1,7 +1,0 @@
-package com.igobrilhante.github.commons
-
-import org.slf4j.{Logger, LoggerFactory}
-
-trait Logging {
-  val logger: Logger = LoggerFactory.getLogger(this.getClass)
-}
