+++ src/main/scala/com/igobrilhante/github/domain/entities/GHCommit.scala
@@ -1,0 +1,5 @@
+package com.igobrilhante.github.domain.entities
+
+case class GHCommit()
+
+object GHCommit {}
