+++ src/main/scala/com/igobrilhante/github/domain/entities/GHContributor.scala
@@ -1,0 +1,5 @@
+package com.igobrilhante.github.domain.entities
+
+case class GHContributor(id: Long, login: String, contributions: Int)
+
+object GHContributor {}
