+++ src/main/scala/com/igobrilhante/github/domain/entities/GHRepository.scala
@@ -1,0 +1,5 @@
+package com.igobrilhante.github.domain.entities
+
+case class GHRepository(id: Long, name: String, `private`: Boolean)
+
+object GHRepository {}
