+++ src/main/scala/com/igobrilhante/github/infraestructure/http/GHSystem.scala
@@ -1,0 +1,11 @@
+package com.igobrilhante.github.infraestructure.http
+
+import akka.actor.typed.Behavior
+import akka.actor.typed.scaladsl.Behaviors
+
+object GHSystem {
+
+  def apply(): Behavior[GHSystem.Command] = Behaviors.empty
+
+  trait Command
+}
